docker/alpr/pyproject.toml
[project]
name = "comma-alpr"
version = "0.1.0"
description = "FastALPR HTTP wrapper for the comma-personal-backend ALPR sidecar."
requires-python = ">=3.12,<3.13"

# Dependencies are pinned to specific versions so the image is
# reproducible across rebuilds. Bump intentionally and re-run the
# alpr-engine-spike harness when changing fast-alpr or onnxruntime --
# both have direct accuracy / latency impact.
dependencies = [
  "fast-alpr==0.1.5",
  "onnxruntime==1.20.1",
  "fastapi==0.115.6",
  "uvicorn[standard]==0.32.1",
  "python-multipart==0.0.20",
  "pillow==11.0.0",
  "numpy==2.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
